Toastmasters Timer page
I made this page to be used as a simple timer app by the Toastmasters club. The screen changes colors to coincide with speech lengths and timing regulations. It also saves all of the speech times recorded and they can be viewed via the "Results" button. Also, the "qualify" field is represented by an emoji. There are multiple levels of success and two levels of failure. If the speech is too short you get 🤔. If the speech is too long you get 🤐. If you qualify but you're within 10 seconds of the max or minimum time after 30 seconds grace is applied then you get 😅. If you are within 15 seconds of the target time in the middle of the speech length ( a 30-second window) you get 😁
 and otherwise you get 😃. 

The page also uses localStorage if available to save the speeches and includes a button to delete these from your local storage.